pyproject.toml
[project]
name = "note-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-deepseek>=1.0.0",
    "load-dotenv>=0.1.0",
]
